requirements.txt
# Core - always installed
python-dotenv
requests
tqdm

# Backend (fastapi or flask)
fastapi
flask
flask-cors
uvicorn[standard]
pydantic

# Worker (Celery + Redis)
celery[redis]
redis

# Scraper (lib)
seleniumbase
beautifulsoup4

# Minimal webapp
streamlit

# AI / Agent SDK
anthropic

# ML (heavy - comment out if not needed)
torch
tensorboard
numpy
onnxruntime
